requirements.txt
streamlit>=1.30
python-chess>=1.999